Move get_signature_schema_from_offered to utils crate with better rustdoc
